requirements.txt
altgraph==0.17.4  
art==6.2     
colorama==0.4.6   
keyboard==0.13.5  
MouseInfo==0.1.3   
numpy==1.26.4  
ordered-set==4.1.0
packaging==24.0
pefile==2023.2.7
pillow==10.3.0
pip==24.0
psutil==5.9.8
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
pymongo==3.11.4
PyMsgBox==1.0.9
pynput==1.7.7
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytweening==1.2.0
pywin32==306
pywin32-ctypes==0.2.2
setuptools==70.0.0
six==1.16.0
termcolor==2.4.0
typing==3.7.4.3
windows-tools.misc==1.0.0
windows-tools.product-key==0.3.2
windows-tools.registry==1.1.0
windows-tools.wmi-queries==1.0.0
WMI==1.5.1
zstandard==0.22.0